requirements.txt
snowflake-connector-python
pandas
scikit-learn
pyyaml